Add files via upload
I applied the necessary codes and configurations to complete the installation of my T-POT using my virtual machine such as Debian Linux on the VirtualBox and a cloud based platform such as Microsoft Azure to create the ultimate multi-Honeypot platform to analyze cyber attacks. It was a great personal achievement!